refactor: config.tsをconfigディレクトリにmv
